# d1/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API REST (LARAVEL 7) IDESA ",
        "description": "API REST construido con Laravel 7 y Documentado con Swagger",
        "contact": {
            "email": "rifarca@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://desafio.idesa.pyaeveapps.com/v2/public"
        }
    ],
    "paths": {
...
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Lista los usuarios",
                "operationId": "29593cb6d735e27309ce747b0c9fbc61",
                "parameters": [
                    {
                        "name": "Authentication",
                        "in": "path",
                        "description": "Authentication Bearer [token] esto va en la cabecera ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "email del usuario",
                                "value": "Authentication Bearer [token]"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Si autentica devuelve un JSON con el ACCESS TOKEN .",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Devuelve un Objeto JSON con la lista de ",
                                        "value": {
                                            "status": "ok",
                                            "message": "Auteticacion de Usuario validado con Exito",
                                            "data": {
                                                "total": 1,
                                                "detail": {
                                                    "id": 1,
                                                    "nombres": "IDESA",
                                                    "correo": "desafio@idesa.com.py",
                                                    "access_token": "3|LdDMJM9HFVi8hqfkKmlP3q7rPUejd4n6Wx1aldWS",
                                                    "type_token": "Bearer",
                                                    "value_token": "Bearer 3|LdDMJM9HFVi8hqfkKmlP3q7rPUejd4n6Wx1aldWS",
                                                    "expiration_token": "2800"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Devuelve un JSON con el error .",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Devuelve un Objeto JSON com el mensaje de error",
                                        "value": {
                                            "status": "Unauthorized",
                                            "message": "No tiene autorizacion para consumir la API"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Modifica/Actualiza datos del Cliente ",
                "operationId": "471739115a8c5082af09e8094c034206",
                "parameters": [
                    {
                        "name": "Authentication",
                        "in": "path",
                        "description": "Authentication Bearer [token] esto va en la cabecera ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "email del usuario",
                                "value": "Authentication Bearer [token]"
                            }
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Authentication Bearer [token] esto va en la cabecera ",
                        "required": true,
                        "schema": {
                            "type": "int"
                        },
                        "examples": {
                            "int": {
                                "summary": "Id del Cliente que se desea modificar",
                                "value": "12345678"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Si autentica devuelve un JSON .",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Devukleve un mensaje de modificacion con exiro",
                                        "value": {
                                            "status": "ok",
                                            "message": "Cliente modificado con Exito"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Devuelve un JSON con el error .",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Devuelve un Objeto JSON com el mensaje de error",
                                        "value": {
                                            "status": "Unauthorized",
                                            "message": "No tiene autorizacion para consumir la API"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crea un Cliente ",
                "operationId": "0f9aa4a97799ea24201a926ec7a0af3d",
                "parameters": [
                    {
                        "name": "Authentication",
                        "in": "path",
                        "description": "Authentication Bearer [token] esto va en la cabecera ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "tokem",
                                "value": "Authentication Bearer [token]"
                            }
                        }
                    },
                    {
                        "name": "nombres",
                        "in": "path",
                        "description": "Nombre del Cliente ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "Nombre del cliente",
                                "value": "Juan"
                            }
                        }
                    },
                    {
                        "name": "apellidos",
                        "in": "path",
                        "description": "Apellido del Cliente ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "Apelloido del cliente",
                                "value": "Perez"
                            }
                        }
                    },
                    {
                        "name": "correo",
                        "in": "path",
                        "description": "email del Cliente ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "email del cliente",
                                "value": "Juan@gmail.com"
                            }
                        }
                    },
                    {
                        "name": "domicilio",
                        "in": "path",
                        "description": "Nombre del Cliente ",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "domicilio del cliente",
                                "value": "Juan ]"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Si autentica devuelve un JSON con el ACCESS TOKEN .",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Devuelve un Objeto JSON con la lista de ",
                                        "value": {
                                            "status": "ok",
                                            "message": "Cliente nuevo creado con Exito!!",
                                            "data": {
                                                "total": 1,
                                                "detail": {
                                                    "nombres": "Fabio",
                                                    "apellidos": "Mendoza",
                                                    "correo": "fabio.mendoza@gmail.com",
                                                    "domicilio": "en algun lugar del meta verso",
                                                    "updated_at": "2023-01-29T22:33:59.000000Z",
                                                    "created_at": "2023-01-29T22:33:59.000000Z",
                                                    "id": 123467
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Devuelve un JSON con el error .",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Devuelve un Objeto JSON com el mensaje de error",
                                        "value": {
                                            "status": "Unauthorized",
                                            "message": "No tiene autorizacion para consumir la API"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "AuthBearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "https",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "AuthBearer": []
        }
    ]
}